+++ bindings/nostr-ffi/src/nips/nip01.rs
@@ -44,6 +44,12 @@ impl From<Coordinate> for nip01::Coordinate {
     }
 }
 
+impl From<&Coordinate> for nip01::Coordinate {
+    fn from(value: &Coordinate) -> Self {
+        value.inner.clone()
+    }
+}
+
 #[uniffi::export]
 impl Coordinate {
     #[uniffi::constructor]